+++ src/library/GameHacks.cpp
@@ -35,7 +35,6 @@ pid_t GameHacks::finalizer_pid = 0;
 uintptr_t GameHacks::executableBase = 0;
 uintptr_t GameHacks::executableEnd = 0;
 uintptr_t GameHacks::unityLoadingThreadAddr = 0;
-std::ptrdiff_t GameHacks::unityLoadingThreadId = 0;
 
 void GameHacks::setUnity()
 {
@@ -76,7 +75,7 @@ void GameHacks::getExecutableMemory()
     executableEnd = reinterpret_cast<uintptr_t>(area.endAddr);
 }
 
-bool GameHacks::isUnityLoadingThread(uintptr_t addr, std::ptrdiff_t routine_id)
+bool GameHacks::isUnityLoadingThread(uintptr_t addr)
 {
     if (!unity) return false;
     
@@ -86,11 +85,10 @@ bool GameHacks::isUnityLoadingThread(uintptr_t addr, std::ptrdiff_t routine_id)
     getExecutableMemory();
     
     if (unityLoadingThreadAddr != 0)
-        return (unityLoadingThreadAddr == addr) && (unityLoadingThreadId == routine_id);
+        return (unityLoadingThreadAddr == addr);
 
     if ((addr >= executableBase) && (addr < executableEnd)) {
         unityLoadingThreadAddr = addr;
-        unityLoadingThreadId = routine_id;
         return true;
     }